+++ src/automapping/detection.py
@@ -1,9 +1,8 @@
 from dataclasses import dataclass, field
-from typing import Sequence
 
 import pandas as pd
 
-from sample_ref import Sample
+from .sample import Sample
 
 
 @dataclass
