--- src/automapping_ref/language_ref.py
@@ -1,10 +1,0 @@
-from enum import Enum
-
-
-class Language(Enum):
-    """
-    An explicitely specified language.
-    """
-
-    GERMAN = "de"
-    ENGLISH = "en"
